requirements.txt
streamlit==1.28.1
requests==2.31.0
websocket-client==1.6.4
python-dotenv==1.0.0
pandas>=2.2.0
numpy>=1.26.0,<2.0
plotly==5.17.0
